Support initializing example code via query param
This is necessary to link out to FHE vs ZKP examples. There's a little
bit of an awkward dance around the initial page load, since the
playground holds configuration state in query params and relies on these
params when navigating routes (e.g. going to /help then back to index
and keeping the users options for debug/release etc.).
Naively adding the example state would mean that such a navigation would
clobber any edits the user made to the code.

Instead, we only update the code based on the example query param _once_
and then the operation is a no-op. After this initial page load, the
only way to change the example code is through the example drop down
menu. In particular, navigating to help and back will _not_ reload the
example code. But, importantly, it will keep the last selected example
in the dropdown selected.